requirements.txt
streamlit
ultralytics
opencv-python
numpy
Pillow
torch
torchvision
torchaudio
